+++ src/firestore/firebase.ts
@@ -7,7 +7,7 @@ let serviceAccount;
 if (process.env.NODE_ENV === 'prod') {
   serviceAccount = JSON.parse(fs.readFileSync(path.resolve(process.env.FIREBASE_CONFIG_PATH!), 'utf8'));
 } else {
-  serviceAccount = JSON.parse(fs.readFileSync(path.resolve(__dirname, '../configs/service_account.json'), 'utf8'));
+  serviceAccount = JSON.parse(fs.readFileSync(path.resolve(__dirname, '../../configs/service_account.json'), 'utf8'));
 }
  
 admin.initializeApp({
